refactor: Marker HTML 가져오는 함수를 다른 파일로 분리
Related to: #38
